+++ crates/connectors/ndc-postgres/src/configuration.rs
@@ -7,6 +7,7 @@ pub mod version2;
 use custom_trait_implementations::RawConfigurationCompat;
 use ndc_sdk::connector;
 use query_engine_metadata::metadata;
+use query_engine_sql::sql::ast::transaction::IsolationLevel;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
@@ -80,6 +81,7 @@ pub struct RuntimeConfiguration {
     pub metadata: metadata::Metadata,
     pub pool_settings: version1::PoolSettings,
     pub connection_uri: String,
+    pub isolation_level: IsolationLevel,
 }
 
 /// Apply the common interpretations on the Configuration API type into an RuntimeConfiguration.
@@ -91,6 +93,7 @@ pub fn as_runtime_configuration(config: &Configuration) -> RuntimeConfiguration 
             connection_uri: match &v1_config.connection_uri {
                 ConnectionUri::Uri(ResolvedSecret(uri)) => uri.clone(),
             },
+            isolation_level: IsolationLevel::default(),
         },
         RawConfiguration::Version2(v2_config) => RuntimeConfiguration {
             metadata: v2_config.metadata.clone(),
@@ -98,6 +101,7 @@ pub fn as_runtime_configuration(config: &Configuration) -> RuntimeConfiguration 
             connection_uri: match &v2_config.connection_uri {
                 ConnectionUri::Uri(ResolvedSecret(uri)) => uri.clone(),
             },
+            isolation_level: v2_config.isolation_level.clone(),
         },
     }
 }