+++ crates/query-engine/sql/src/sql/helpers.rs
@@ -567,18 +567,22 @@ pub const VARIABLES_OBJECT_PLACEHOLDER: &str = "%VARIABLES_OBJECT_PLACEHOLDER";
 /// SQL field name to be used for ordering results with multiple variable sets.
 pub const VARIABLE_ORDER_FIELD: &str = "%variable_order";
 
-pub fn mutation_begin() -> Vec<string::Statement> {
+pub fn begin(
+    isolation_level: &transaction::IsolationLevel,
+    transaction_mode: transaction::TransactionMode,
+) -> Vec<string::Statement> {
     vec![{
         let mut sql = string::SQL::new();
         transaction::Begin {
-            isolation_level: transaction::IsolationLevel::ReadCommitedReadWrite,
+            isolation_level: isolation_level.clone(),
+            transaction_mode,
         }
         .to_sql(&mut sql);
         string::Statement(sql)
     }]
 }
 
-pub fn mutation_end() -> Vec<string::Statement> {
+pub fn commit() -> Vec<string::Statement> {
     vec![{
         let mut sql = string::SQL::new();
         transaction::Commit {}.to_sql(&mut sql);