+++ crates/query-engine/translation/tests/common/mod.rs
@@ -3,8 +3,15 @@ use std::fs;
 use query_engine_sql::sql;
 use query_engine_translation::translation;
 
-/// Translate a query to SQL and compare against the snapshot.
 pub fn test_translation(testname: &str) -> Result<String, translation::error::Error> {
+    test_query_translation(&None, testname)
+}
+
+/// Translate a query to SQL and compare against the snapshot.
+pub fn test_query_translation(
+    isolation_level: &Option<sql::ast::transaction::IsolationLevel>,
+    testname: &str,
+) -> Result<String, translation::error::Error> {
     let tables = serde_json::from_str(
         fs::read_to_string(format!("tests/goldenfiles/{}/tables.json", testname))
             .unwrap()
@@ -18,27 +25,62 @@ pub fn test_translation(testname: &str) -> Result<String, translation::error::Er
     )
     .unwrap();
 
-    let plan = translation::query::translate(&tables, request)?;
-    let plan = plan.query;
-    let query = plan.query_sql();
-    let params: Vec<(usize, &sql::string::Param)> = query
-        .params
-        .iter()
-        .enumerate()
-        .map(|(i, p)| (i + 1, p))
-        .collect();
+    let plan = translation::query::translate(
+        &tables,
+        &isolation_level
+            .clone()
+            .unwrap_or(sql::ast::transaction::IsolationLevel::default()),
+        request,
+    )?;
+
+    let mut sqls: Vec<String> = vec![];
+
+    if isolation_level.is_some() {
+        for stmt in plan.pre {
+            let pretty = sqlformat::format(
+                &stmt.0.sql,
+                &sqlformat::QueryParams::None,
+                sqlformat::FormatOptions::default(),
+            );
+            sqls.push(pretty);
+        }
+    }
+
+    let query = plan.query.query_sql();
 
     let pretty = sqlformat::format(
         &query.sql,
         &sqlformat::QueryParams::None,
         sqlformat::FormatOptions::default(),
     );
+    sqls.push(pretty);
 
-    Ok(format!("{}\n\n{:?}", pretty, params))
+    let param: Vec<(usize, sql::string::Param)> = query
+        .params
+        .into_iter()
+        .enumerate()
+        .map(|(i, p)| (i + 1, p))
+        .collect();
+
+    if isolation_level.is_some() {
+        for stmt in plan.post {
+            let pretty = sqlformat::format(
+                &stmt.0.sql,
+                &sqlformat::QueryParams::None,
+                sqlformat::FormatOptions::default(),
+            );
+            sqls.push(pretty);
+        }
+    }
+
+    Ok(format!("{}\n\n{:?}", sqls.join(";\n\n"), param))
 }
 
 /// Translate a mutation to SQL and compare against the snapshot.
-pub fn test_mutation_translation(testname: &str) -> Result<String, translation::error::Error> {
+pub fn test_mutation_translation(
+    isolation_level: &Option<sql::ast::transaction::IsolationLevel>,
+    testname: &str,
+) -> Result<String, translation::error::Error> {
     let tables = serde_json::from_str(
         fs::read_to_string(format!(
             "tests/goldenfiles/mutations/{}/tables.json",
@@ -70,7 +112,12 @@ pub fn test_mutation_translation(testname: &str) -> Result<String, translation::
         })
         .collect::<Result<Vec<_>, translation::error::Error>>()?;
 
-    let plan = sql::execution_plan::simple_mutations_execution_plan(mutations);
+    let plan = sql::execution_plan::simple_mutations_execution_plan(
+        &isolation_level
+            .clone()
+            .unwrap_or(sql::ast::transaction::IsolationLevel::default()),
+        mutations,
+    );
     let mut sqls: Vec<String> = vec![];
     let mut params: Vec<Vec<(usize, sql::string::Param)>> = vec![];