+++ crates/query-engine/translation/tests/tests.rs
@@ -236,7 +236,21 @@ mod mutations {
 
     #[test]
     fn simple() {
-        let result = common::test_mutation_translation("simple").unwrap();
+        let result = common::test_mutation_translation(&None, "simple").unwrap();
+        insta::assert_snapshot!(result);
+    }
+}
+
+mod transaction {
+    use crate::common;
+
+    #[test]
+    fn select_with_limit() {
+        let result = common::test_query_translation(
+            &Some(query_engine_sql::sql::ast::transaction::IsolationLevel::Serializable),
+            "select_with_limit",
+        )
+        .unwrap();
         insta::assert_snapshot!(result);
     }
 }