+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Query transactions now run in read-only mode, and allow transaction isolation level configuration. ([#209](https://github.com/hasura/ndc-postgres/pull/209))
 - Support variables as arguments to native queries ([#211](https://github.com/hasura/ndc-postgres/pull/211))
 - Introduce version 2 of connector deployment configuration. ([#208](https://github.com/hasura/ndc-postgres/pull/208))
 - Support array types ([#191](https://github.com/hasura/ndc-postgres/pull/191), ...)